A negative size_t was being converted into an int64_t which resulted in the wrong value.

The problem is a negative unsigned value is still unsigned, and then it was stored in an int64_t. So the result was a large positive value instead of a small negative one. Fix is to use the right type.

Related work items: #886

+++ lotus/core/providers/cpu/math/element_wise_ops.cc
@@ -263,7 +263,7 @@ struct BroadcastIterator {
   }
 
   std::vector<int64_t> counters_;
-  std::vector<int64_t> deltas_;
+  std::vector<ptrdiff_t> deltas_;
   std::vector<int64_t> counts_;
   size_t count_{1};  // Running total count of entries in tensor, used while building up the entries
